ck_project/requirements.txt
Django==2.1.7
django-filter==2.1.0
djangorestframework==3.9.2
Markdown==3.0.1
pytz==2018.9
Pillow==5.4.1
psycopg2-binary
